optimize:move cmd to alpha
