+++ cmd/sealer/cmd/alpha/alpha.go
@@ -1,0 +1,36 @@
+// Copyright © 2021 Alibaba Group Holding Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package alpha
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// NewCmdAlpha returns "sealer alpha" command.
+func NewCmdAlpha() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "alpha",
+		Short: "sealer experimental sub-commands",
+	}
+
+	cmd.AddCommand(NewPruneCmd())
+	cmd.AddCommand(NewDebugCmd())
+	cmd.AddCommand(NewExecCmd())
+	cmd.AddCommand(NewMergeCmd())
+	cmd.AddCommand(NewUpgradeCmd())
+	cmd.AddCommand(NewGenCmd())
+	cmd.AddCommand(NewCertCmd())
+	return cmd
+}
